src/requirements_old.txt
accelerate
peft
bitsandbytes
transformers
python-dotenv
huggingface-hub
protobuf
sentencepiece
faiss-cpu
sentence-transformers